requirements.txt
streamlit 
pandas 
plotly.express 
geopandas 
folium
folium_static
linear
requests

